mass reformatting and cleanup of compiler warnings

+++ common/extrabiomes/handlers/SaplingBonemealEventHandler.java
@@ -6,33 +6,36 @@
 
 package extrabiomes.handlers;
 
-import extrabiomes.blocks.BlockCustomSapling;
-import extrabiomes.blocks.BlockNewSapling;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockSapling;
 import net.minecraftforge.event.Event.Result;
 import net.minecraftforge.event.ForgeSubscribe;
 import net.minecraftforge.event.entity.player.BonemealEvent;
+import extrabiomes.blocks.BlockCustomSapling;
 
-public class SaplingBonemealEventHandler {
-
+public class SaplingBonemealEventHandler
+{
+    
     private final BlockCustomSapling sapling;
-
-    public SaplingBonemealEventHandler(BlockCustomSapling sapling) {
+    
+    public SaplingBonemealEventHandler(BlockCustomSapling sapling)
+    {
         this.sapling = sapling;
     }
-
+    
     @ForgeSubscribe
-    public void onBonemealEvent(BonemealEvent e) {
-        if (e.getResult() == Result.DEFAULT && e.ID == sapling.blockID) {
-        	if (!e.world.isRemote) {
-                if ((double)e.world.rand.nextFloat() < 0.45D) {
-                	sapling.markOrGrowMarked(e.world, e.X, e.Y, e.Z, e.world.rand);
+    public void onBonemealEvent(BonemealEvent e)
+    {
+        if (e.getResult() == Result.DEFAULT && e.ID == sapling.blockID)
+        {
+            if (!e.world.isRemote)
+            {
+                if (e.world.rand.nextFloat() < 0.45D)
+                {
+                    sapling.markOrGrowMarked(e.world, e.X, e.Y, e.Z, e.world.rand);
                 }
                 
                 e.setResult(Result.ALLOW);
             }
         }
     }
-
+    
 }
